requirements.txt
python-dateutil>=2.8.2
typing-extensions>=4.5.0
openai>=1.0.0
python-dotenv>=1.0.0
langchain
chromadb
pypdf
tiktoken
pydantic
pandas
streamlit>=1.31.0
plotly>=5.19.0